New user functional, added Confirm to the delete button that shows an alert of Cancel and ok

+++ public/lukasApp.js
@@ -5,16 +5,18 @@ let posts = [] ;
 
 
 // get posts
-/* Jag gör en fetch på
+/* Jag gör en fetch 
 */
+// Gets the data from the DB and displays it on the screen
+// While it is doing that, the loading screen shows until the data is fully recivied
 const getPosts = async () => {
   openLoadScreen()
     const res = await fetch(`${rootURL}getposts`); 
     const data = await res.json();
     posts = data.posts
     document.querySelector("tbody").innerHTML = posts
-    .map((post, index) =>   
-        `
+    
+    .map((post, index) => `
         <tbody>
               <tr>
                 <td>${index += 1}</td>
@@ -38,21 +40,22 @@ const getPosts = async () => {
     closeLoadScreen()
 }
 
-
+//Delete function that handles the id has been passed through and asking the user to delete or not 
+// With the if statement and confirm.
+//if the user is deleted then getPosts function runs again to display the new data
 const deleteUser = async (id) => {
-
   openLoadScreen()
-
+  if(confirm("Do you really want to delete this user?")){
   const ressponce = await fetch(`${rootURL}deletepost/${id}`, {
         method: "delete"
-      
     })
-
+    
     getPosts()
+    
+  }else{
     closeLoadScreen()
-      
+  }
 } 
-
 
 //------------------------- LOAD SCREEN START --------------------------------//
 
@@ -66,6 +69,5 @@ const closeLoadScreen = () => {
 
 //------------------------- LOAD SCREEN END --------------------------------//
 
-
-
+//When the window is finished loading, the getPosts fucntion triggers and displays the data.
 window.addEventListener("load", getPosts())
